Create Basic code area and perimeter of rectangle
It is a simple code.
Easy to understand how the code works
The labels are 'Enter the length', 'Enter the Breadth' and 'Answer' for each Entry box.
The commands are 'area' and 'perimeter'. If area button clicked, the area will calculate and answer displayed on 'Answer' Entry. If perimeter button clicked, the perimeter will calculate and answer displayed on 'Answer' Entry.
It is easy to learn.